Fixed some bugs in the mm and finished adding rsdt searching

+++ lib/common/common.h
@@ -26,8 +26,8 @@
 // Alignment
 ////////////////////////////////////////////////////////////////////////////
 
-#define ALIGN_DOWN(n, a) ((n) & ~((a) - 1))
-#define ALIGN_UP(n, a) ALIGN_DOWN((n) + (a) - 1, (a))
+#define ALIGN_DOWN(n, a) (((uint64_t)n) & ~((a) - 1))
+#define ALIGN_UP(n, a) ALIGN_DOWN(((uint64_t)n) + (a) - 1, (a))
 #define ALIGN_DOWN_PTR(p, a) ((void *)ALIGN_DOWN((uintptr_t)(p), (a)))
 #define ALIGN_UP_PTR(p, a) ((void *)ALIGN_UP((uintptr_t)(p), (a)))
